+++ packages/core/src/ruleset/formats.ts
@@ -1,10 +1,10 @@
-import type { Format } from '../format';
+import type { Format } from './format';
 
 function printFormat(format: Format): string {
   return format.displayName ?? format.name;
 }
 
-export class FormatsSet<T extends Format = Format> extends Set<T> {
+export class Formats<T extends Format = Format> extends Set<T> {
   public toJSON(): string[] {
     return Array.from(this).map(printFormat);
   }