+++ packages/core/src/ruleset/validation/__tests__/__fixtures__/valid-flat-ruleset.ts
@@ -1,0 +1,12 @@
+import { truthy } from '@stoplight/spectral-functions';
+
+export default {
+  rules: {
+    'valid-rule': {
+      given: '$.info',
+      then: {
+        function: truthy,
+      },
+    },
+  },
+};